Strip accents from accented characters
After the string map, we pass the result to transform.String which uses
normalization magic[1] to move the modifiers (accents etc) from the
characters.  The intermediate variable slug is used so that
transform.String's n and err results can be ignored.

I've included a test, checking that:
    Árvíztűrő Tükörfúrógép
is transformed to:
    arvizturo-tukorfurogep

1: https://blog.golang.org/normalization#TOC_10.